fix(cli): use the repo's u32 ranged value_parser for --max-concurrency
clap::value_parser!(usize) resolves to _AnonymousValueParser, which has no
.range(). cook.rs:521 already uses value_parser!(u32).range(1..); match it and
convert at the three consumption sites.

+++ crates/homeboy-cli/src/commands/agent_task/args/definitions/fanout.rs
@@ -93,10 +93,10 @@ pub struct AgentTaskFanoutCookBatchArgs {
     /// it are reported as `concurrency` in the batch result.
     #[arg(
         long = "max-concurrency",
-        value_parser = clap::value_parser!(usize).range(1..),
+        value_parser = clap::value_parser!(u32).range(1..),
         value_name = "N"
     )]
-    pub max_concurrency: Option<usize>,
+    pub max_concurrency: Option<u32>,
     /// Wall-clock budget, in seconds, for the whole batch — every child, every
     /// attempt, and every gate.
     ///
@@ -177,10 +177,10 @@ pub struct AgentTaskFanoutRunPlanArgs {
     /// `fanout cook-batch --max-concurrency`.
     #[arg(
         long = "max-concurrency",
-        value_parser = clap::value_parser!(usize).range(1..),
+        value_parser = clap::value_parser!(u32).range(1..),
         value_name = "N"
     )]
-    pub max_concurrency: Option<usize>,
+    pub max_concurrency: Option<u32>,
     /// Wall-clock budget, in seconds, for the whole batch. See
     /// `fanout cook-batch --max-duration`.
     #[arg(
